athena_system/requirements.txt
langchain
langchain-community
langchain-openai
duckduckgo-search
python-dotenv
pydantic
termcolor
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
chromadb
ai-core-sdk
